requirements.txt
flask
flask-restful
flask_sqlalchemy
python-dotenv
flask_jwt_extended
fastavro